refactor: UserRoleClientResponse 로 이름 변경
