+++ src/main/java/com/mople/dto/client/UserRoleClientResponse.java
@@ -9,32 +9,32 @@ import lombok.Builder;
 import java.util.List;
 
 @Builder
-public record UserClientResponse(
+public record UserRoleClientResponse(
         Long userId,
         String nickname,
         String image,
         UserRole role
 ) {
-    public static List<UserClientResponse> ofParticipants(List<PlanParticipant> participants, Long hostId, Long creatorId) {
+    public static List<UserRoleClientResponse> ofParticipants(List<PlanParticipant> participants, Long hostId, Long creatorId) {
         return participants.stream()
                 .map(p -> ofUser(p.getUser(), hostId, creatorId))
                 .toList();
     }
 
-    public static List<UserClientResponse> ofAutoCompleteUsers(List<MeetMember> members, Long hostId, Long creatorId) {
+    public static List<UserRoleClientResponse> ofAutoCompleteUsers(List<MeetMember> members, Long hostId, Long creatorId) {
         return members.stream()
                 .map(m -> ofUser(m.getUser(), hostId, creatorId))
                 .toList();
     }
 
-    public static List<UserClientResponse> ofMembers(List<MeetMember> members, Long hostId) {
+    public static List<UserRoleClientResponse> ofMembers(List<MeetMember> members, Long hostId) {
         return members.stream()
                 .map(m -> ofUser(m.getUser(), hostId))
                 .toList();
     }
 
-    private static UserClientResponse ofUser(User user, Long hostId, Long creatorId) {
-        return UserClientResponse.builder()
+    private static UserRoleClientResponse ofUser(User user, Long hostId, Long creatorId) {
+        return UserRoleClientResponse.builder()
                 .userId(user.getId())
                 .nickname(user.getNickname())
                 .image(user.getProfileImg())
@@ -42,8 +42,8 @@ public record UserClientResponse(
                 .build();
     }
 
-    private static UserClientResponse ofUser(User user, Long hostId) {
-        return UserClientResponse.builder()
+    private static UserRoleClientResponse ofUser(User user, Long hostId) {
+        return UserRoleClientResponse.builder()
                 .userId(user.getId())
                 .nickname(user.getNickname())
                 .image(user.getProfileImg())
