+++ src/main/java/com/mople/meet/controller/MeetController.java
@@ -2,7 +2,7 @@ package com.mople.meet.controller;
 
 import com.mople.core.annotation.auth.SignUser;
 import com.mople.dto.client.MeetClientResponse;
-import com.mople.dto.client.UserClientResponse;
+import com.mople.dto.client.UserRoleClientResponse;
 import com.mople.dto.request.meet.MeetCreateRequest;
 import com.mople.dto.request.meet.MeetUpdateRequest;
 import com.mople.dto.request.pagination.CursorPageRequest;
@@ -83,7 +83,7 @@ public class MeetController {
             description = "모임 유저 목록을 조회합니다."
     )
     @GetMapping("/members/{meetId}")
-    public ResponseEntity<FlatCursorPageResponse<UserClientResponse>> getMeetMembers(
+    public ResponseEntity<FlatCursorPageResponse<UserRoleClientResponse>> getMeetMembers(
             @Parameter(hidden = true) @SignUser AuthUserRequest user,
             @PathVariable Long meetId,
             @ParameterObject @Valid CursorPageRequest request
